+++ config/atompark-sms-channel.php
@@ -1,0 +1,6 @@
+<?php
+
+// config for Andriichuk/AtomParkSmsChannel
+return [
+
+];