+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `laravel-atompark-sms-channel` will be documented in this file.
+
+## 0.1.0 - 2026-03-11
+
+- Initial release of the AtomPark SMS notification channel
+- Adds `AtomParkChannel` for sending SMS via AtomPark
+- Provides `Sms` value object and `AtomParkClient` HTTP client